requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=1.5.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.0
xlsxwriter>=3.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
